requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.17.0
MarkupSafe==1.1.1
numpy==1.19.2
pandas==1.1.2
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
Werkzeug==1.0.1
wincertstore==0.2
